Improved HCA encoding and decoding with Encryption
Made an improvised port of VGAudio, and also added in the better HCA encrypting and decrypting methods, now they work in C.
I also changed the PCM code a bit, as well as fixing bugs in the IO handlers.

I do have an improved ADX code ready, but would need me to test it to push it eventually. TODO for now.

+++ CriCodecs/BitWriter.cpp
@@ -4,7 +4,7 @@
 class BitWriter{
 public:
     unsigned char* buffer;
-    unsigned int len; // This would give maximum file buffer to be 512 megabytes (0xFFFFFFFF * 8).
+    unsigned long long len; // This would give maximum file buffer to be 512 megabytes (0xFFFFFFFF * 8).
     unsigned int pos;
 
     int get_remaining(){
